Switch to CUDA driver API and make it possible to link against CUEW
Main idea of this change is to make it possible to produce portable
OpenSubdiv builds by loading CUDA symbols from it's dynamic library
on runtime. This way it's possible to make the same executable file
to run on computers with and without CUDA toolkit installed.

It's not possible to load symbols from runtime CUDA library (well,
it's possible but it wouldn't work as expected) because of the way
how kernels are registering (they're using some pre-load magic
which is rather tricky to control and make work reliably).

So the idea is to switch to driver API which is much easier to make
loaded on runtime. In most cases it's pretty straighforward changes
in functions calls, but here're some notes about changed behavior:

- Changes around cudaGLSetGLDevice() are not fully tested because
  of the lack of dual-GPU computer here. But the idea here is to
  pass device index to CudaComputeController, so it can use whatever
  device is chosen by external user.

  By default it'll be device #0, but examples application will set
  it to GPU with maximun FLOPS.

- cudaDeviceReset is gone now, it was used to clean up resources on
  exit, and now it should be done by destructor of CUDA compute
  controller class.

To make portable builds CUEW library is to be used:
  https://github.com/Nazg-Gul/cuew

Untested changes:
- Dual-GPU setups
- Maya plugins
- DirectX backends