Fix compilation error of DirectX backend
Some portions of code weren't ported to a
driver api still. Also needed to tweak linker
flags for dxViewer in order to make it built
on my laptop,

I'm not so much sure we can get rid of the
call of cudaD3D11SetDirect3DDevice() but
seems it's all working now.

NOTE: cuew wouldn't really work with DX.